Demo Update
Added
- ladders
- steam suit
cleaned up some movement scripts and modified prefabs to be more robust. bugs in movement will be hammered out later
