+++ Steam Shift/Assets/Player/ENGINEER.prefab
@@ -13,6 +13,7 @@ GameObject:
   - component: {fileID: 3728902827064375603}
   - component: {fileID: 3728902827064375600}
   - component: {fileID: 3728902827064375601}
+  - component: {fileID: -5090300580192869781}
   m_Layer: 7
   m_Name: ENGINEER
   m_TagString: Untagged
@@ -100,14 +101,18 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   engineer: {fileID: 3728902827064375603}
+  go: {fileID: 3728902827064375606}
   rb: {fileID: 0}
-  spriteRenderer: {fileID: 0}
+  sr: {fileID: 0}
+  colliders: []
   pipe: {fileID: 1553060239226734580, guid: 2c406ab6fdafc684587c54c3176bff2b, type: 3}
+  numPipes: 0
   speed: 0
   jumpForce: 0
+  isClimbing: 0
   grounded: 0
   jumping: 0
-  floorCollider: {fileID: 3728902827064375601}
+  floorCollider: {fileID: -5090300580192869781}
   floorFilter:
     useTriggers: 0
     useLayerMask: 0
@@ -141,7 +146,7 @@ Rigidbody2D:
   m_Material: {fileID: 6200000, guid: b470303d0dd2a4c4b874c3d3d28dfe18, type: 2}
   m_Interpolate: 0
   m_SleepingMode: 1
-  m_CollisionDetection: 0
+  m_CollisionDetection: 1
   m_Constraints: 4
 --- !u!61 &3728902827064375601
 BoxCollider2D:
@@ -153,7 +158,7 @@ BoxCollider2D:
   m_Enabled: 1
   m_Density: 1
   m_Material: {fileID: 6200000, guid: b470303d0dd2a4c4b874c3d3d28dfe18, type: 2}
-  m_IsTrigger: 0
+  m_IsTrigger: 1
   m_UsedByEffector: 0
   m_UsedByComposite: 0
   m_Offset: {x: 0, y: 0}
@@ -168,4 +173,30 @@ BoxCollider2D:
   m_AutoTiling: 0
   serializedVersion: 2
   m_Size: {x: 1, y: 1}
+  m_EdgeRadius: 0
+--- !u!61 &-5090300580192869781
+BoxCollider2D:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 3728902827064375606}
+  m_Enabled: 1
+  m_Density: 1
+  m_Material: {fileID: 6200000, guid: b470303d0dd2a4c4b874c3d3d28dfe18, type: 2}
+  m_IsTrigger: 0
+  m_UsedByEffector: 0
+  m_UsedByComposite: 0
+  m_Offset: {x: 0, y: -0.25}
+  m_SpriteTilingProperty:
+    border: {x: 0, y: 0, z: 0, w: 0}
+    pivot: {x: 0.5, y: 0.5}
+    oldSize: {x: 1, y: 1}
+    newSize: {x: 1, y: 1}
+    adaptiveTilingThreshold: 0.5
+    drawMode: 0
+    adaptiveTiling: 0
+  m_AutoTiling: 0
+  serializedVersion: 2
+  m_Size: {x: 0.9, y: 0.5}
   m_EdgeRadius: 0
