+++ Steam Shift/Assets/Scenes/PlatformMockup.unity
@@ -653,23 +653,13 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   steam: {fileID: 1550717479}
+  go: {fileID: 0}
   rb: {fileID: 0}
-  spriteRenderer: {fileID: 0}
+  sr: {fileID: 0}
+  colliders: []
   speed: 1
-  steamFilter:
-    useTriggers: 0
-    useLayerMask: 0
-    useDepth: 0
-    useOutsideDepth: 0
-    useNormalAngle: 0
-    useOutsideNormalAngle: 0
-    layerMask:
-      serializedVersion: 2
-      m_Bits: 0
-    minDepth: 0
-    maxDepth: 0
-    minNormalAngle: 0
-    maxNormalAngle: 0
+  isClimbing: 0
+  isSuit: 0
 --- !u!61 &1550717481
 BoxCollider2D:
   m_ObjectHideFlags: 0
@@ -908,10 +898,15 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   engineer: {fileID: 1670551949}
+  go: {fileID: 0}
   rb: {fileID: 0}
-  spriteRenderer: {fileID: 0}
+  sr: {fileID: 0}
+  colliders: []
+  pipe: {fileID: 0}
+  numPipes: 0
   speed: 0
   jumpForce: 0
+  isClimbing: 0
   grounded: 0
   jumping: 0
   floorCollider: {fileID: 1670551950}
