+++ Steam Shift/Assets/Scenes/SampleLevel1.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: abbff78e828634c40aae5f079961ed0d
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
