+++ Steam Shift/Assets/Scripts/GridTiles/Goal.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e1705b90c3b29064b86dd1716b7a6376
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
