+++ Steam Shift/Assets/Scripts/GridTiles/Ladder.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 806422989ee61674ca8d63505214c323
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
